test(identity): DistributorFactory must hash PAN to raw bytes, not hex
distributors.pan_hash is BINARY(32) and every production writer passes
$binary = true to hash('sha256', ...). The factory did not, so it wrote the
64-char hex digest. sqlite does not enforce column widths, so this only ever
surfaced on MySQL as error 1406 — where it errored every test that built a
distributor through the factory.

Unblocking it exposed three assertions that the crash had been hiding:
MySQL returns SUM() as a string through PDO while sqlite returns an int, so
comparing sum('amount_paise') to an int with toBe() failed on the type. Adds
the (int) cast the rest of the suite already uses; the amounts were correct.

Net effect on the two affected files: 32 errored, 0 assertions -> 26 passed.

+++ app/tests/Modules/Compensation/MentorshipBonusServiceTest.php
@@ -98,7 +98,7 @@ it('credits the sponsor with slab points × the day\'s pooled point value (slab 
     expect($mb->status)->toBe('credited');
 
     // Sponsor wallet credited with gross.
-    expect(WalletLedgerEntry::where('distributor_id', $sponsor->id)->sum('amount_paise'))->toBe(525_000);
+    expect((int) WalletLedgerEntry::where('distributor_id', $sponsor->id)->sum('amount_paise'))->toBe(525_000);
 });
 
 it('pays each slab its own points at one shared value (slab 3 → 15 pts; slab 7 → 3 pts, both @ ₹250)', function () {
@@ -118,7 +118,7 @@ it('pays each slab its own points at one shared value (slab 3 → 15 pts; slab 7
     expect($mbA->mb_gross_paise)->toBe(375_000);
     expect($mbB->msb_points)->toBe(3);
     expect($mbB->mb_gross_paise)->toBe(75_000);
-    expect(WalletLedgerEntry::where('distributor_id', $sponsor->id)->sum('amount_paise'))->toBe(450_000);
+    expect((int) WalletLedgerEntry::where('distributor_id', $sponsor->id)->sum('amount_paise'))->toBe(450_000);
 });
 
 it('prices every credit at the day\'s frozen pool value, whatever the slab', function () {
@@ -232,7 +232,7 @@ it('is idempotent — calling twice for the same cutoff does not double-credit',
     expect($second)->not->toBeNull();
     expect(MentorshipBonusResult::count())->toBe(1);
     expect(WalletLedgerEntry::where('distributor_id', $sponsor->id)->count())->toBe(1);
-    expect(WalletLedgerEntry::where('distributor_id', $sponsor->id)->sum('amount_paise'))->toBe(525_000);
+    expect((int) WalletLedgerEntry::where('distributor_id', $sponsor->id)->sum('amount_paise'))->toBe(525_000);
 });
 
 it('blocks MB credit when sponsor personal BV is below the minimum threshold', function () {
